Several multiprocessing bug fixed, refactored init of database, refactored recommender class
Now multiprocessing is fully working, thanks to queue and manager from multiprocessing.
Kittens is run by default as a multiprocess application because recommender has been heavily modified to allow efficient multiprocess.
Kittens launcher reflect the changes.
Database init no more contains the retrieval of csv since this would conflict with the multiprocess manager.
All csv retrieval is done via staticmethod of helper class
Ultimely, since the dataset is changed, a lot of work had to be done on the database to make the switch, and is in constant refactor