fixed testing performance issue, added tui, added new recommendation method
Fixed a performance issue when computing tests (Test Mode)

Added TUI (run.sh) and changed kittens accordingly. The new TUI is asking
the same parameter of kittens but in visual form. To format output for the progress
bar some changes have been made to the output of the engine. Insert logging
to restore debug info need to be done.

Added a new recommendation method (make_tags_recommendations) based on the tags which
features in the playlists as a whole. Some modification was also required in the DB